updating from source master

+++ CI/install-dependencies-osx.sh
@@ -30,7 +30,9 @@ brew update
 #Base OBS Deps and ccache
 brew install jack speexdsp ccache mbedtls freetype fdk-aac
 brew install https://gist.githubusercontent.com/DDRBoxman/9c7a2b08933166f4b61ed9a44b242609/raw/ef4de6c587c6bd7f50210eccd5bd51ff08e6de13/qt.rb
-brew unlink swig
+if [ -d "$(brew --cellar)/swig" ]; then
+    brew unlink swig
+fi
 brew install https://gist.githubusercontent.com/DDRBoxman/4cada55c51803a2f963fa40ce55c9d3e/raw/572c67e908bfbc1bcb8c476ea77ea3935133f5b5/swig.rb
 
 pip install dmgbuild
@@ -40,8 +42,8 @@ ccache -s || echo "CCache is not available."
 
 # Fetch and untar prebuilt OBS deps that are compatible with older versions of OSX
 hr "Downloading OBS deps"
-wget --quiet --retry-connrefused --waitretry=1 https://github.com/obsproject/obs-deps/releases/download/2020-4-16/osx-deps-2020-04-16.tar.gz
-tar -xf ./osx-deps-2020-04-16.tar.gz -C /tmp
+wget --quiet --retry-connrefused --waitretry=1 https://github.com/obsproject/obs-deps/releases/download/2020-04-24/osx-deps-2020-04-24.tar.gz
+tar -xf ./osx-deps-2020-04-24.tar.gz -C /tmp
 
 # Fetch vlc codebase
 hr "Downloading VLC repo"
